+++ src/GeneGraph.tsx
@@ -1,11 +1,14 @@
-import React, { useState, useEffect, useMemo, createContext, useContext, useReducer } from 'react';
+import React, { useState, useEffect, useMemo, createContext, useContext, useCallback } from 'react';
 import { useExpand } from './store/store';
-import { ReactFlow, Background, Controls, MiniMap, useNodesState, useEdgesState, Handle, ReactFlowProvider } from 'reactflow';
+import { ReactFlow, Background, Controls, MiniMap, useNodesState, useEdgesState, useReactFlow, ReactFlowProvider, Panel } from 'reactflow';
 import 'reactflow/dist/style.css';
 import { nodeTypes } from "./NodeTypes";
 import FloatingEdge from './EdgeType';
 import FloatingConnectionLine from './FloatingConnectionLine';
 import { SidebarFilterList } from './SidebarFilterList';
+
+import ELK from 'elkjs/lib/elk.bundled.js';
+
 
 export const NodesContext = React.createContext(null)
 
@@ -19,6 +22,51 @@ type GeneGraphProps = {
   setIds;
 };
 
+
+const elk = new ELK();
+
+const useLayoutedElements = () => {
+  const { getNodes, setNodes, getEdges, fitView } = useReactFlow();
+  const defaultOptions = {
+    'elk.algorithm': 'org.eclipse.elk.force',
+    'elk.layered.spacing.nodeNodeBetweenLayers': 100,
+    'elk.spacing.nodeNode': 80,
+  };
+
+  const getLayoutedElements = useCallback(() => {
+    const layoutOptions = { ...defaultOptions };
+    const graph = {
+      id: 'root',
+      layoutOptions: layoutOptions,
+      children: getNodes(),
+      edges: getEdges(),
+    };
+
+    elk.layout(graph as any).then(({ children }) => {
+
+      // By mutating the children in-place we saves ourselves from creating a
+      // needless copy of the nodes array.
+      children.forEach((node) => {
+        (node as any).position = { x: node.x, y: node.y };
+      });
+
+      setNodes(children as any);
+
+      window.requestAnimationFrame(() => {
+        fitView({
+          maxZoom: 15,
+          minZoom: 0.1,
+          duration: 5000,
+          nodes: getNodes()
+        });
+      });
+    });
+  }, []);
+
+  return { getLayoutedElements };
+};
+
+
 // GeneGraph component
 export function GeneGraph(props: GeneGraphProps) {
   let geneIds = props.geneID;
@@ -27,43 +75,81 @@ export function GeneGraph(props: GeneGraphProps) {
   // state for the nodes and edges
   const [nodes, setNodes, onNodesChange] = useNodesState([]);
   const [edges, setEdges, onEdgesChange] = useEdgesState([]);
+  const { getLayoutedElements } = useLayoutedElements();
+  const [curNodes, setCurNodes] = useState(0);
+
 
 
   // get all genes that are connected to the first node
   let { data: graph } = useExpand({
-      geneIds: geneIds,
+    geneIds: geneIds,
     limit: 1000,
   });
 
+  const { getNodes, fitView, getEdges } = useReactFlow();
+
+  useEffect(() => {
+    if (getNodes() != null && getNodes()?.length != curNodes && getNodes()[0]["width"] != null) {
+      setCurNodes(getNodes().length);
+      getLayoutedElements();
+    }
+
+    //TODO 
+    // //fit only if number of nodes has changed
+    // // if (getNodes().length != currentNodes) {
+    // //   setCurrentNodes(getNodes().length);
+    // //   fitView({
+    // //     maxZoom: 15,
+    // //     minZoom: 0.1,
+    // //     duration: 5000,
+    // //     nodes: getNodes()
+    // //   });
+    // }
+
+    window.requestAnimationFrame(() => {
+      fitView({
+        maxZoom: 15,
+        minZoom: 0.1,
+        duration: 5000,
+        nodes: getNodes()
+      });
+    });
+
+
+  }, [getNodes(), getEdges()]);
+
   useMemo(() => {
+    setCurNodes(0);
     (document as any).startViewTransition(() => {
-      currentNodes = graph?.nodes.map((node,index)=>{
-        return{
-          id:node.id,
-          position:{
-            x:(node.position.x+1)*(window.innerWidth/2),
-            y:(node.position.y+1)*(window.innerHeight/2)
+
+      currentNodes = graph?.nodes.map((node, index) => {
+        return {
+          id: node.id,
+          position: {
+            x: (node.position.x + 1) * (window.innerWidth / 2),
+            y: (node.position.y + 1) * (window.innerHeight / 2)
           },
-          data:{
-            label:node.symbol == "nan" ? node.id : node.symbol,
-            displayProps:{
-              fullname:node.name,
-              synonyms:node.synonyms,
-              entrezId:node.entrezId,
-              label:node.symbol == "nan" ? node.id : node.symbol,
-              summary:node.summary,
+          data: {
+            label: node.symbol == "nan" ? node.id : node.symbol,
+            displayProps: {
+              fullname: node.name,
+              synonyms: node.synonyms,
+              entrezId: node.entrezId,
+              label: node.symbol == "nan" ? node.id : node.symbol,
+              summary: node.summary,
             },
             children: node.children,
             parents: node.parents,
             isRoot: props.geneID.includes(node.id) ? true : false,
-            type:node.type,
+            type: node.type,
             onExpand: exp,
             onCollapse: coll
           },
           type: "node"
-        }});
+        }
+      });
       setNodes(currentNodes)
-      
+
       setEdges(
         graph?.edges.map((edge) => ({
           id: edge.id,
@@ -74,20 +160,22 @@ export function GeneGraph(props: GeneGraphProps) {
       );
 
     })
+
   }, [graph]);
 
-  function exp(id: string){
-    if(!geneIds.includes(id)){
-      geneIds = ([...geneIds,id])
+  function exp(id: string) {
+    if (!geneIds.includes(id)) {
+      geneIds = ([...geneIds, id])
       props.setIds(geneIds)
     }
   }
 
-  function coll(id: string, children: [string]){
+  function coll(id: string, children: [string]) {
     geneIds = geneIds.filter(geneId => geneId != id)
     currentNodes.forEach((child) => {
-      child.data.parents = child.data.parents.filter((parent: string) => parent != id)})
-      
+      child.data.parents = child.data.parents.filter((parent: string) => parent != id)
+    })
+
     var removeChildren = currentNodes.filter(node => children.includes(node.id))
     removeChildren = removeChildren.filter(node => node.data?.parents.length == 0)
     currentNodes = currentNodes.filter(node => !removeChildren.includes(node))
@@ -95,36 +183,28 @@ export function GeneGraph(props: GeneGraphProps) {
   }
 
   return (
-    <>
-      <div style={{ height: '90%', width: '100%', display: 'flex' }}>
+    <div style={{ height: '100%', width: '100%', display: 'flex' }}>
+      <ReactFlow
+        nodes={nodes}
+        edges={edges}
+        nodeTypes={nodeTypes}
+        edgeTypes={edgeTypes}
 
-        <ReactFlowProvider>
-          <div style={{ height: '100%', width: '77%' }}>
-            <ReactFlow
-              nodes={nodes}
-              edges={edges}
-              nodeTypes={nodeTypes}
-              edgeTypes={edgeTypes}
-              
-              onNodesChange={onNodesChange}
-              onEdgesChange={onEdgesChange}
-              connectionLineComponent={FloatingConnectionLine}
-              fitView 
-            >
-              <Background />
-              <Controls />
-              <MiniMap />
-            </ReactFlow>
-          </div>
-          <NodesContext.Provider value={{nodes: nodes, setNodes: setNodes}}>
-            <SidebarFilterList />
-          </NodesContext.Provider>
-          
-        </ReactFlowProvider>
+        minZoom={0.1}
+        maxZoom={10}
 
-      </div>
-
-    </>
-
+        onNodesChange={onNodesChange}
+        onEdgesChange={onEdgesChange}
+        connectionLineComponent={FloatingConnectionLine}
+        fitView
+      >
+        <Background />
+        <Controls />
+        <MiniMap />
+      </ReactFlow >
+      <NodesContext.Provider value={{ nodes: nodes, setNodes: setNodes }}>
+        <SidebarFilterList />
+      </NodesContext.Provider>
+    </div>
   );
 }
